🚧 WIP refactor and improve naming
